Minor Changes in UI

+++ example/lib/data/home/now_playing_repo.dart
@@ -59,7 +59,7 @@ class NowPlayResult {
         original_language: json['original_language'],
         original_title: json['original_title'],
         overview: json['overview'],
-        popularity: json['popularity'],
+        popularity: json['popularity'].toDouble(),
         poster_path: json['poster_path'],
         release_date: json['release_date'],
         title: json['title'],
